adds chart help subchart
