+++ claim-gui/src/components/VestingChart/ChartHelp/style.module.css
@@ -1,0 +1,70 @@
+.container {
+  position: absolute;
+  top: -18px;
+  right: -215px;
+  width: 250px;
+  z-index: 1000;
+}
+
+.container .info {
+  padding-left: 2em;
+}
+
+.allocationChart {
+  width: 100%;
+  border-radius: 4px;
+  border: 1px solid;
+  background-color: #f0ebde;
+  overflow: hidden;
+}
+
+.pieChart path {
+  filter: brightness(1);
+  transition: all 0.1s;
+  cursor: crosshair;
+  stroke: black;
+  stroke-width: 1;
+  stroke-linejoin: round;
+}
+
+path.focus {
+  filter: brightness(0.8);
+}
+
+.labels {
+  stroke-width: 0;
+  fill: black;
+  font-size: 24px;
+  text-anchor: middle;
+}
+
+.labels text {
+  transform-box: fill-box;
+  transform-origin: center;
+}
+
+.hoverLabel {
+  pointer-events: none;
+  display: flex;
+  align-items: center;
+  gap: 5px;
+  background: ghostwhite;
+  border: 0.5px solid;
+  border-radius: 4px;
+  padding: 4px 4px;
+  position: absolute;
+  bottom: 12px;
+  left: calc(1em + 8px);
+  opacity: 0;
+}
+
+.hoverLabel.active {
+  opacity: 1;
+}
+
+.hoverLabel p {
+  margin: 0;
+  padding-top: 2px;
+  font-size: 0.8em;
+  line-height: 1em;
+}
