+++ claim-gui/src/components/VestingInfo/style.module.css
@@ -12,71 +12,12 @@
   text-decoration: underline;
 }
 
-.vestingChart {
-  width: 100%;
-  border-radius: 4px;
-  border: 1px solid;
-  background-color: #f0ebde;
-  overflow: hidden;
+.statusContainer {
+  position: absolute;
+  bottom: -4px;
+  right: calc(0.5em + 8px);
 }
 
-.pieChart path {
-  filter: brightness(1);
-  transition: all 0.1s;
-  cursor: crosshair;
-  stroke: black;
-  stroke-width: 0.5;
-  stroke-linejoin: round;
-}
-
-.marks {
-  stroke: black;
-  stroke-width: 0.5;
-  stroke-linejoin: round;
-  fill: none;
-}
-
-.labels {
-  stroke-width: 0;
-  fill: black;
-  font-size: 12px;
-}
-
-.labels text {
-  transform-box: fill-box;
-  transform-origin: center;
-}
-
-.labels tspan {
-  text-anchor: middle;
-}
-
-path.elseAllocation {
-  transform: scale(0.95);
-  fill: #cbffdc;
-}
-
-path.elseAllocation.focus {
-  filter: brightness(0.96);
-}
-
-path.gnosisAllocation {
-  fill: #caefff;
-}
-
-path.gnosisAllocation.focus {
-  filter: brightness(0.92);
-}
-
-path.gnosisVested {
-  fill: #3ea077;
-}
-
-path.gnosisVested.focus {
-  filter: brightness(0.88);
-}
-
-.hoverLabel,
 .connectionStatus {
   pointer-events: none;
   display: flex;
@@ -86,28 +27,6 @@ path.gnosisVested.focus {
   border: 0.5px solid;
   border-radius: 4px;
   padding: 4px 4px;
-}
-
-.hoverLabel {
-  position: absolute;
-  bottom: -4px;
-  left: calc(0.5em + 8px);
-  opacity: 0;
-}
-
-.hoverLabel.active {
-  opacity: 1;
-}
-
-.hoverLabel p {
-  margin: 0;
-  line-height: 1em;
-}
-
-.statusContainer {
-  position: absolute;
-  bottom: -4px;
-  right: calc(0.5em + 8px);
 }
 
 .indicator {
